+++ web-common/src/lib/time/relative-time.ts
@@ -1,0 +1,24 @@
+/**
+ * Get relative time for recent dates (e.g., "2m ago", "1h ago")
+ */
+export function getRelativeTime(dateString: string): string {
+  if (!dateString) return "";
+
+  try {
+    const date = new Date(dateString);
+    const now = new Date();
+    const diffInMinutes = Math.floor(
+      (now.getTime() - date.getTime()) / (1000 * 60),
+    );
+
+    if (diffInMinutes < 1) return "now";
+    if (diffInMinutes < 60) return `${diffInMinutes}m ago`;
+
+    const diffInHours = Math.floor(diffInMinutes / 60);
+    if (diffInHours < 24) return `${diffInHours}h ago`;
+
+    return "";
+  } catch {
+    return "";
+  }
+}
